chore : 객체 역할 작게 유지하기
