+++ src/main/java/com/baseballgame/rule/BallChecker.java
@@ -2,6 +2,7 @@ package com.baseballgame.rule;
 
 public class BallChecker {
     StrikeChecker strikeChecker = new StrikeChecker();
+    NumberComparator numberComparator = new NumberComparator();
 
     public int getBallResult(int[] userNumbers, Integer[] answer) {
         return subtractStrikeFromSum(userNumbers, answer);
@@ -27,15 +28,9 @@ public class BallChecker {
     private int countUserNumberInAnswer(int userNumber, Integer[] answer) {
         int countUserNumberInAnswer = 0;
         for (int i = 0; i < answer.length; i++) {
-            countUserNumberInAnswer += compareNumber(answer[i], userNumber);
+            countUserNumberInAnswer += numberComparator.compareNumber(answer[i], userNumber);
         }
         return countUserNumberInAnswer;
     }
 
-    private int compareNumber(int answerNumber, int userNumber) {
-        if (answerNumber == userNumber) {
-            return 1;
-        }
-        return 0;
-    }
 }
